requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
docling==2.15.0
chromadb==0.4.22
sentence-transformers==2.3.1
google-generativeai==0.3.2
pypdf==4.0.1
python-dotenv==1.0.1
numpy<2.0.0
